[AAE-77] moved ProcessExtensionServiceIT to boot-starter
